Move containers into a separate directory. Correct include files accordingly
