+++ StPicoEvent/StPicoBTOWHit.h
