+++ StPicoEvent/StPicoEmcTrigger.h
