+++ StPicoEvent/StPicoMtdHit.h
