+++ StPicoEvent/StPicoMtdTrigger.cxx
